run: new module
move here the code for running a new container
