+++ Makefile.am
@@ -36,6 +36,7 @@ bwrap_oci_SOURCES = \
 	list.c \
 	safe-read-write.c \
 	util.c \
+	run.c \
 	subugidmap.c \
 	$(NULL)
 
@@ -51,7 +52,7 @@ TESTS = tests/bind.conf tests/cwd.conf tests/env.conf tests/hostname.conf \
 TEST_EXTENSIONS = .conf
 CONF_LOG_COMPILER = $(top_srcdir)/tests/tests-runner
 
-EXTRA_DIST = autogen.sh bwrap-oci.xml safe-read-write.h util.h list.h
+EXTRA_DIST = autogen.sh bwrap-oci.xml safe-read-write.h util.h list.h run.h
 
 EXTRA_DIST += $(TESTS:.conf=.conf.expected)
 EXTRA_DIST += $(TESTS:.conf=.conf.command)
